Refactored a lot of animation code into a separate js file
